refactor(DeleteUserUseCase): It checks if the user does not exist and throws an error.

+++ src/modules/users/use-cases/DeleteUserUseCase.ts
@@ -1,5 +1,6 @@
 import User from "../../../entities/User";
 import { IUserRepository } from "../../../interfaces/IUserRepository";
+import UseCaseError from "../../../interfaces/UseCaseError";
 
 class DeleteUserUseCase {
   private userRepository: IUserRepository;
@@ -8,6 +9,15 @@ class DeleteUserUseCase {
     this.userRepository = userRepository;
   }
   async execute(userId: number): Promise<void> {
+    const existingUser = await this.userRepository.findById(userId);
+
+    if (!existingUser) {
+      throw <UseCaseError>{
+        message: "User not found",
+        errorType: "NOT_FOUND",
+      };
+    }
+
     await this.userRepository.deleteById(userId);
   }
 }
